requirements.txt
pandas
openpyxl
python-dotenv
openai
click
PyMuPDF 